requirements.txt
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 3.3.1
selenium==3.141.0
urllib3==1.25.6
